+++ src/Exception/ResourceNotFound.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Scriptotek\Alma\Exception;
+
+
+class ResourceNotFound extends ClientException
+{
+}